Refactor GetFirmwareVersionsCommandExecutor to make use of DlmsHelperService. Renames GetFirmwareVersionExecutor to GetFirmwareVersionsExecutor to better match its function.
