+++ osgp-protocol-adapter-dlms/src/main/java/org/osgp/adapter/protocol/dlms/application/services/ConfigurationService.java
@@ -15,7 +15,7 @@ import org.openmuc.jdlms.MethodResultCode;
 import org.openmuc.jdlms.SecurityUtils.KeyId;
 import org.osgp.adapter.protocol.dlms.application.models.ProtocolMeterInfo;
 import org.osgp.adapter.protocol.dlms.domain.commands.GetAdministrativeStatusCommandExecutor;
-import org.osgp.adapter.protocol.dlms.domain.commands.GetFirmwareVersionCommandExecutor;
+import org.osgp.adapter.protocol.dlms.domain.commands.GetFirmwareVersionsCommandExecutor;
 import org.osgp.adapter.protocol.dlms.domain.commands.GetPushSetupSmsCommandExecutor;
 import org.osgp.adapter.protocol.dlms.domain.commands.ReplaceKeyCommandExecutor;
 import org.osgp.adapter.protocol.dlms.domain.commands.SetActivityCalendarCommandActivationExecutor;
@@ -100,7 +100,7 @@ public class ConfigurationService {
     private GetAdministrativeStatusCommandExecutor getAdministrativeStatusCommandExecutor;
 
     @Autowired
-    private GetFirmwareVersionCommandExecutor getFirmwareVersionCommandExecutor;
+    private GetFirmwareVersionsCommandExecutor getFirmwareVersionCommandExecutor;
 
     @Autowired
     private ReplaceKeyCommandExecutor replaceKeyCommandExecutor;
